# projeto_Renovacao/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "API de Gerenciamento de Empresas, Cursos e Demandas",
      "description": "API para gerenciamento de empresas, instituições, cursos, professores e demandas, com funcionalidades CRUD e integração de processos educacionais e empresariais.",
      "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
    "/empresas/{id}":{
      "put":{
        "tags":["Empresas"],
        "summary": "Atualiza os dados de uma empresa existente",
        "description": "Este endpoint atualiza os dados de uma empresa com as informações fornecidas.",
        "operationId": "atualizarEmpresa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da empresa a ser atualizada.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nome": {
                      "type": "string",
                      "description": "Nome da empresa a ser atualizado"
                    },
                    "email": {
                      "type": "string",
                      "description": "Email da empresa"
                    },
                    "cep": {
                      "type": "string",
                      "description": "CEP da empresa"
                    },
                    "pais": {
                "type": "string",
                "description": "País da empresa"
              },  "contato": {
                "type": "string",
                "description": "Contato da empresa"
              },
              "cnpj": {
                "type": "string",
                "description": "CNPJ da empresa"
              },
              "senha": {
                "type": "string",
                "description": "Senha da empresa"
              }

                  }
                }
                
            }
          }
        },
        "responses":{
          "204":{
            "description": "Empresa atualizada com sucesso, sem corpo na resposta."
  
          },
          "500": {
            "description": "Erro no servidor interno."
          }
        }
        
      },

      "patch":{
        "tags":["Empresas"],
        "summary": "Atualiza a senha de uma empresa existente",
        "description": "Este endpoint atualiza a senha de uma empresa com as informações fornecidas.",
        "operationId": "patchMudarSenha",
        "parameters": [
          {
              "name": "id",
              "in": "path",
              "description": "ID da empresa cuja senha será atualizada.",
              "required": true,
              "schema": {
                  "type": "integer"
              }
          }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nova_senha": {
                    "type": "string",
                    "description": "Nova senha para a empresa"
                }
                }
              }
              }
            }
      },
      "responses": {
        "204":{
          "description": "Senha da empresa atualizada com sucesso, sem corpo na resposta."

        }
        , "500": {
          "description": "Erro no servidor interno."
      }
      }
      },
      "delete":{
        "tags":["Empresas"],
        "summary": "Deleta uma empresa existente",
        "description": "Este endpoint deleta uma empresa com o pelo id fornecido.",
        "operationId": "deletarEmpresa",
        "parameters": [
          {
              "name": "id",
              "in": "path",
              "description": "ID da empresa que será deletada.",
              "required": true,
              "schema": {
                  "type": "integer"
              }
          }
      ],
      "responses": {
        "204":{
          "description": "Empresa deletada com sucesso, sem corpo na resposta."

        },
         "500": {
          "description": "Erro no servidor interno."
      }
      }
      }
     
    },
...
      "/instituicoes/{id}":{
        "put":{
          "tags":["Instituição"],
          "summary": "Atualiza os dados de uma instituição existente",
          "description": "Este endpoint atualiza os dados de uma Instituição com as informações fornecidas.",
          "operationId": "atualizarEmpresa",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID de uma Instituição a ser atualizada.",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "nome": {
                        "type": "string",
                        "description": "Nome da instituiçãoa ser atualizado"
                      },
                      "email": {
                        "type": "string",
                        "description": "Email da instituição"
                      },
                      "cep": {
                        "type": "string",
                        "description": "CEP da instituição"
                      },
                      "pais": {
                  "type": "string",
                  "description": "País da instituição"
                },  "contato": {
                  "type": "string",
                  "description": "Contato da instituição"
                },
                "cnpj": {
                  "type": "string",
                  "description": "CNPJ da instituição"
                },
                "senha": {
                  "type": "string",
                  "description": "Senha da instituição"
                }
  
                    }
                  }
                  
              }
            }
          },
          "responses":{
            "204":{
              "description": "Instituição atualizada com sucesso, sem corpo na resposta."
    
            },
            "500": {
              "description": "Erro no servidor interno."
            }
          }
          
        },
  
        "patch":{
          "tags":["Instituição"],
          "summary": "Atualiza a senha de uma instituição existente",
          "description": "Este endpoint atualiza a senha de uma instituição com as informações fornecidas.",
          "operationId": "patchMudarSenha",
          "parameters": [
            {
                "name": "id",
                "in": "path",
                "description": "ID da instituição cuja senha será atualizada.",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nova_senha": {
                      "type": "string",
                      "description": "Nova senha para a instituição"
                  }
                  }
                }
                }
              }
        },
        "responses": {
          "204":{
            "description": "Senha da instituição atualizada com sucesso, sem corpo na resposta."
  
          }
          , "500": {
            "description": "Erro no servidor interno."
        }
        }
        },
        "delete":{
          "tags":["Instituição"],
          "summary": "Deleta uma instituição existente",
          "description": "Este endpoint deleta uma instituição com o pelo id fornecido.",
          "operationId": "deletarInstituição",
          "parameters": [
            {
                "name": "id",
                "in": "path",
                "description": "ID da empresa que será deletada.",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        ],
        "responses": {
          "204":{
            "description": "Empresa deletada com sucesso, sem corpo na resposta."
  
          },
           "500": {
            "description": "Erro no servidor interno."
        }
        }
        }
       
      },
...
  },
...
}